format: 마크다운 기능 함수 파일 분리
