+++ pages/write/listener.js
@@ -1,0 +1,4 @@
+/**
+ * @description write binding 함수
+ */
+export const writeEventBinding = () => {};
