+++ pages/write/render.js
@@ -1,27 +1,5 @@
 import { getNode } from '../../lib/utils/getNode.js';
-
-/**
- * 카테고리, 제목 태그 생성
- * @param {string} label
- * @returns {HTMLDivElement} 태그 생성
- */
-const createInputTags = label => {
-  const placeholder = label === 'Category' ? '카테고리를 입력해 주세요.' : '제목을 입력해 주세요.';
-
-  return `
-      <h1 class="label">${label}</h1>
-      <input type="text" placeholder="${placeholder}" />
-    `;
-};
-
-/**
- *  write, preview 버튼 생성
- * @param {string} label
- * @returns {HTMLDivElement} 태그 생성
- */
-const createTabButton = label => {
-  return `<button id="write-${label}-button" type="button" class="${label === 'Write' ? 'is-button-active' : ''}">${label}</button>`;
-};
+import { createInputTags, createTabButton } from './create.js';
 
 /**
  * 카테고리 태그 렌더링
